backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
sqlalchemy
mysqlclient
python-dotenv
passlib
python-jose[cryptography]